felix/bpf: disabled CTLB workaround for host and nodeports
Packets from the host are intercepted at the first device they hit. If
it is towards the same hos, it is the loopback device, otherwise it is
the main device on the cluster network. At that point we do the DNAT -
select the backed pod for the nodeport

Then we either use FIB to redirect the packet back to the same node (if
a pod on the same node was selected) or we redirect the packet to the
bpfin.cali iface to turn the packet around into the host address
namespace that can route the packet. If an overlay is used, we SNAT the
packet so that it has the tunnel's source and we redirect it to the
tunnel.

+++ felix/bpf/ut/pol_prog_test.go
@@ -1585,7 +1585,7 @@ func (p packet) StateIn() state.State {
 		preNATDstAddr = p.preNATDstAddr
 		preNATDstPort = p.preNATDstPort
 	}
-	flags := uint8(0)
+	flags := uint64(0)
 	if p.fromHostFlag {
 		flags |= polprog.FlagSrcIsHost
 	}
